space

+++ Src/main.c
@@ -153,7 +153,7 @@ char motor_poles = 14;
 char drag_brake_strength = 10;		// Drag Brake Power
 char sine_mode_changeover_thottle_level = 5;	// Sine Startup Range
 char sine_mode_changeover_mutliplier = 15;
-int sine_mode_changeover = 5 * 15;
+short sine_mode_changeover = 5 * 15;
 
 char USE_HALL_SENSOR = 0;
